% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsp}
\alias{lsp}
\title{Least Squares Problems in Surveying}
\format{
\describe{
 \item{Binary}{ra component of the right-hand-side as binary form}
 \item{Continuous}{ra component of the right-hand-side as continuous form}
}
}
\usage{
lsp
}
\description{
One matrices from the least-squares solution of problems in surveying that were used by Michael Saunders and Chris Paige in the testing of LSQR
}
\references{
Koenker, R and Ng, P. (2002). SparseM: A Sparse Matrix Package for R,
http://www.econ.uiuc.edu/~roger/research/home.html
Matrix Market,
https://math.nist.gov/MatrixMarket/data/Harwell-Boeing/lsq/lsq.html
}
\keyword{datasets}
