% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{PIC}
\alias{PIC}
\title{Generic method for PIC (fractional form)}
\usage{
PIC(object, n, const1 = 1.5, const2 = 0.8, ...)
}
\arguments{
\item{object}{an object with a PIC method.}

\item{n}{the number of observations.}

\item{const1}{the constant for DF.}

\item{const2}{the constant for IF.}

\item{...}{additional arguments passed to methods.}
}
\description{
Generic method for PIC (fractional form)
}
