% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{PIC_inner}
\alias{PIC_inner}
\title{Internal PIC function - not meant to be directly accessed by users.}
\usage{
PIC_inner(loss, DF, IF, mn, const1 = 1.5, const2 = 0.8)
}
\arguments{
\item{loss}{the squared error loss.}

\item{DF}{degrees of freedom.}

\item{IF}{inflation factor.}

\item{mn}{the number of observations (univariate), or the number of responses times the number of observations (multivariate).}

\item{const1}{the constant for DF.}

\item{const2}{the constant for IF.}
}
\description{
Internal PIC function - not meant to be directly accessed by users.
}
