% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TISP_cpp.R
\name{print.SPAS}
\alias{print.SPAS}
\title{Print method for SPAS}
\usage{
\method{print}{SPAS}(x, ...)
}
\arguments{
\item{x}{a \code{SPAS} object.}

\item{...}{not used.}
}
\description{
Print method for SPAS
}
