% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TISP_cpp.R
\name{print.STAVE}
\alias{print.STAVE}
\title{Print method for STAVE}
\usage{
\method{print}{STAVE}(x, ...)
}
\arguments{
\item{x}{a \code{STAVE} object.}

\item{...}{not used.}
}
\description{
Print method for STAVE
}
