% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_linreg.R
\name{sparse_linreg}
\alias{sparse_linreg}
\title{sparse_linreg}
\usage{
sparse_linreg(X, y)
}
\arguments{
\item{X}{covariate matrix X}

\item{y}{outcome vector y}
}
\value{
This function would return a data frame

\code{Estimate}: the estimated coefficients

\code{SE}: standard errors

\code{T}:  T statistics

\code{p_value}:  p_value of the corresponding statistics
}
\description{
Conduct linear regression on sparse matrix
}
\examples{
#One dataset from LSQ
data(lsp)
X <- lsp[,3:714]
y_C <- lsp[,1]
lin_fit <- sparse_linreg(X,y_C)
}
