% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_logreg.R
\name{sparse_logreg}
\alias{sparse_logreg}
\title{sparse_logreg}
\usage{
sparse_logreg(outcome_column, design_columns, data_frame)
}
\arguments{
\item{outcome_column}{the column name of the binary outcome}

\item{design_columns}{the column names corresponding to the covariates}

\item{data_frame}{the name of the data frame}
}
\value{
A matrix with same number of rows the length of the design_columns argument (corresponding to β estimates) and 4 columns.

\code{Estimate}: the estimated coefficients

\code{SE}: standard errors

\code{T}:  T statistics

\code{p_value}:  p_value of the corresponding statistics
}
\description{
Conduct logistic regression on sparse matrix
}
\examples{
# simulate mateix
set.seed(12032020)
sample.space<-c(rep(0,40),seq(1,10))
mat<-matrix(sample(sample.space,n*m,replace=T),nrow=50000,ncol=100)
mat<-cbind(sample(c(0,1),n,replace=T),mat)
colnames(mat)<-c('environment',paste0('gene',seq(1,100)))
text<-as.data.frame(mat)
log_fit<-sparse_logreg(outcome_column = 'environment',design_columns = paste0('gene',seq(1,15)),data_frame = test)
}
