% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAS.log.likelihood.R
\name{SPAS.likelihood.star.DM}
\alias{SPAS.likelihood.star.DM}
\title{Score, likelihood, and related functions for fitting Stratified Petersen model}
\usage{
SPAS.likelihood.star.DM(
  Est.Star,
  rowDM,
  colDM,
  thetaDM,
  rawdata,
  returnnegll = FALSE,
  conditional = TRUE,
  fixed.N = 0,
  mytrace = FALSE
)
}
\description{
These functions are used internally to fit the Stratified Petersen models and are not normally used.
Most of these functions are for the legacy version that does not use TMB and so likely can be eventually deleted.
}
\keyword{internal}
