% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAS.print.model.R
\name{SPAS.print.model}
\alias{SPAS.print.model}
\title{Print the results from a fit of a Stratifed-Petersen (SP) model}
\usage{
SPAS.print.model(x)
}
\arguments{
\item{x}{A result from the model fitting. See \code{SPAS.fit.model}.}
}
\value{
A report to the console.
}
\description{
This function makes a report of the results of the model fitting
.
}
\examples{
Refer to the vignettes.

}
