% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TISP_cpp.R
\name{SPAS}
\alias{SPAS}
\title{Run iterative threshodling for Semi-Parametric Affinity Based Selection (SPAS)}
\usage{
SPAS(X_affinity, y_affinity, lambda, y_sc = NULL, beta_init = NULL,
  eta = 1e-04, accelerate = F, line_search = T, step_size = NULL,
  ls_beta = 1.5, ls_eps = 1e-10, ls_max_iter = 100,
  thresh_type = c("quantile_ridge", "quantile", "soft", "hard",
  "hard_ridge"), tol_type = c("beta", "loss"), max_iter = 2000,
  eps = 1e-05, ls_type = c("forward", "backward"))
}
\arguments{
\item{X_affinity}{a symmetric matrix of affinities between predictors.}

\item{y_affinity}{a vector of affinities between each predictor and the response.}

\item{lambda}{the tuning paramemter. Either a non-negative real number for soft or hard thresholding or a non-negative integer for quantile thresholding.}

\item{y_sc}{a positive real number representing the scaling value for \code{y}. Default is 1.}

\item{beta_init}{an initial value for the beta vector. Default is the zero vector.}

\item{eta}{parameter for ridge type thresholding.}

\item{accelerate}{use over-relaxation to accelerate the optimization.}

\item{line_search}{use line search.}

\item{step_size}{the step-size. The default is the inverse spectral norm of \code{X_affinity} when \code{ls_type} is "forward", or 1 when \code{ls_type} is "backward".}

\item{ls_beta}{by how much should the step-size increase or decrease in each line search iteration?}

\item{ls_eps}{you can miss the line search check by this much.}

\item{ls_max_iter}{the maximum number of line search iterations.}

\item{thresh_type}{the type of thresholding.}

\item{tol_type}{the type of tolerance checking.}

\item{max_iter}{the maximum number of iterations.}

\item{eps}{the maximum value of the tolerance before declaring convergence.}

\item{ls_type}{either backtracking (backward) or forward tracking (forward) line search.}
}
\value{
An objecty of type \code{SPAS} with
\item{\code{beta_opt}}{optimal beta.}
\item{\code{iter}}{number of iterations.}
\item{\code{step_size}}{}
\item{\code{eta}}{}
\item{\code{loss_final}}{final loss value.}
\item{\code{loss_vals}}{vector of loss values at each iteration.}
\item{\code{converged}}{a boolean.}
\item{\code{tot_time}}{total time for iterative thresholding.}
\item{\code{nz_patt}}{the indices of estimated non-zero predictors.}
\item{\code{lambda}}{}
\item{\code{J}}{the estimated cardinality.}
\item{\code{IF}}{the inflation factor (for tuning).}
\item{\code{DF}}{the degrees of freedom (for tuning).}
}
\description{
Run iterative threshodling for Semi-Parametric Affinity Based Selection (SPAS)
}
