% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_to_cor.R
\name{zstat_to_cor}
\alias{zstat_to_cor}
\title{Convert z statistics to correlations.}
\usage{
zstat_to_cor(zstat, n)
}
\arguments{
\item{zstat}{numeric vector of z statistics.}

\item{n}{number of observations.}
}
\description{
Convert z statistics to correlations.
}
