% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAS.misc.functions.r
\name{logit}
\alias{logit}
\title{Helper functions for this package (logit, expit)}
\usage{
logit(p)
}
\arguments{
\item{p}{A proportion.}

\item{theta}{A logit}
}
\description{
These functions are helper functions for the SPAS package and not normally accessed by the user
}
\details{
\code{logit} - compute the logit of a proportion
\code{expit} - compute a proportion from a logit
}
\keyword{internal}
