% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TISP_cpp.R
\name{npn_cor}
\alias{npn_cor}
\title{Create the SKEPTIC estimates for Spearman's rho and Kendall's tau.}
\usage{
npn_cor(x, y = NULL, method = c("spearman", "kendall"))
}
\description{
Create the SKEPTIC estimates for Spearman's rho and Kendall's tau.
}
